+++ rando/core/views.py
@@ -1,4 +1,5 @@
 import os
+import mimetypes
 
 from django.conf import settings
 from django.views.static import serve as static_serve
@@ -9,6 +10,10 @@ def fileserve(request, path):
     Rewrite URLs to use current language as folder root prefix.
     TODO: Could be done with ``mod_rewrite`` at deployment.
     """
+
+    if '.geojson' not in mimetypes.types_map:
+        mimetypes.add_type('application/json', '.geojson')
+
     path = path[1:] if path.startswith('/') else path
     if not os.path.exists(os.path.join(settings.INPUT_DATA_ROOT, path)):
         path = os.path.join(request.LANGUAGE_CODE, path)